cleaning code, ready for push
